lib: move Merge<Tree> impls from merge module to tree module
The `Merge<Tree>` methods and `to_tree_merge()` need `Store` and
`Tree`, which would be a layering violation in the backend module, so
move them to the tree module instead.
